move transformer basis out
