+++ src/nemos/basis/__init__.py
@@ -1,6 +1,7 @@
 from ._basis import AdditiveBasis, Basis, MultiplicativeBasis
 from ._raised_cosine_basis import RaisedCosineBasisLinear, RaisedCosineBasisLog
 from ._spline_basis import BSplineBasis
+from ._transformer_basis import TransformerBasis
 from .basis import (
     ConvBSpline,
     ConvCyclicBSpline,
